Extract FileDropZone
