+++ packages/protocolbeat/src/view/NodeView.tsx
@@ -1,5 +1,5 @@
 import clsx from 'clsx'
-import { useCallback, useEffect, useRef } from 'react'
+import { useCallback, useRef } from 'react'
 
 import type { Node } from '../store/State'
 import { useStore } from '../store/store'
@@ -11,7 +11,6 @@ export interface NodeViewProps {
   node: Node
   selected: boolean
   discovered: boolean
-  onHideNode: (nodeId: string) => void
 }
 
 export function NodeView(props: NodeViewProps) {
@@ -20,27 +19,6 @@ export function NodeView(props: NodeViewProps) {
   const updateNodeLocations = useStore((state) => state.updateNodeLocations)
   // Using ref instead of inline event listener
   // to prevent side-menu from flashing on hidden node
-  const hideRef = useRef<HTMLButtonElement>(null)
-  const onHide = useCallback(
-    (e: MouseEvent) => {
-      e.preventDefault()
-      e.stopPropagation()
-      props.onHideNode(props.node.simpleNode.id)
-    },
-    [props.onHideNode, props.node.simpleNode.id],
-  )
-
-  useEffect(() => {
-    if (hideRef.current) {
-      // The `true` argument here signifies that this listener is for the capture phase
-      // actively prevent store-attached events from firing
-      hideRef.current.addEventListener('mousedown', onHide, true)
-    }
-
-    return () => {
-      hideRef.current?.removeEventListener('mousedown', onHide, true)
-    }
-  }, [])
 
   const onDoubleClick = useCallback(() => {
     if (!ref.current) {
@@ -83,9 +61,6 @@ export function NodeView(props: NodeViewProps) {
         )}
       >
         <div className="truncate">{props.node.simpleNode.name}</div>
-        <div className="flex items-center justify-center gap-2">
-          <button ref={hideRef}>👁️</button>
-        </div>
       </div>
       {props.node.fields.map(({ name, connection }, i) => (
         <div className="relative" key={i}>
