+++ packages/protocolbeat/src/view/Sidebar.tsx
@@ -28,42 +28,36 @@ function SidebarForSingleNode({ node }: { node: SimpleNode }) {
       : undefined
 
   return (
-    <div className="row-span-2 overflow-y-auto bg-white p-2 drop-shadow-xl">
-      <div className="flex flex-col gap-2">
-        <h2 className="font-bold text-xl">{humanReadableName}</h2>
-        <p className="text-gray-500 text-sm">
+    <div className="flex w-[400px] flex-col gap-2 overflow-y-auto bg-white p-2 drop-shadow-xl">
+      <h2 className="font-bold text-xl">{humanReadableName}</h2>
+      <p className="text-gray-500 text-sm">
+        <a
+          href={etherscanLink}
+          target="_blank"
+          title="Etherscan"
+          rel="noreferrer"
+        >
+          {address}
+        </a>{' '}
+        {sourceLink ? (
           <a
-            href={etherscanLink}
+            href={sourceLink}
+            className="font-bold"
             target="_blank"
-            title="Etherscan"
+            title="DethCode"
             rel="noreferrer"
           >
-            {address}
-          </a>{' '}
-          {sourceLink ? (
-            <a
-              href={sourceLink}
-              className="font-bold"
-              target="_blank"
-              title="DethCode"
-              rel="noreferrer"
-            >
-              📜
-            </a>
-          ) : null}
-        </p>
+            📜
+          </a>
+        ) : null}
+      </p>
 
-        <p className="text-gray-500">Details:</p>
-        <pre className="overflow-auto text-sm">
-          <code>
-            {JSON.stringify(
-              node.type !== 'Unknown' ? node.data : null,
-              null,
-              2,
-            )}
-          </code>
-        </pre>
-      </div>
+      <p className="text-gray-500">Details:</p>
+      <pre className="overflow-auto text-sm">
+        <code>
+          {JSON.stringify(node.type !== 'Unknown' ? node.data : null, null, 2)}
+        </code>
+      </pre>
     </div>
   )
 }
